requirements.txt
streamlit>=1.28.0
python-dotenv==1.0.1
openai==1.3.0
requests==2.31.0
stripe==5.4.0
setuptools>=65.0.0
Pillow>=9.0.0
googlemaps
reportlab>=4.0.0
